move freeTile to LevleUtils
